+++ statistics/histogram.go
@@ -19,16 +19,13 @@ import (
 	"fmt"
 	"math"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 	"unsafe"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/charset"
-	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/parser/terror"
 	"github.com/pingcap/tidb/sessionctx"
@@ -40,7 +37,6 @@ import (
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/collate"
 	"github.com/pingcap/tidb/util/logutil"
-	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tipb/go-tipb"
 	"github.com/twmb/murmur3"
@@ -145,80 +141,6 @@ func (hg *Histogram) MemoryUsage() (sum int64) {
 	}
 	sum = EmptyHistogramSize + hg.Bounds.MemoryUsage() + int64(cap(hg.Buckets))*EmptyBucketSize + int64(cap(hg.scalars))*EmptyScalarSize
 	return
-}
-
-// AvgColSize is the average column size of the histogram. These sizes are derived from function `encode`
-// and `Datum::ConvertTo`, so we need to update them if those 2 functions are changed.
-func (c *Column) AvgColSize(count int64, isKey bool) float64 {
-	if count == 0 {
-		return 0
-	}
-	// Note that, if the handle column is encoded as value, instead of key, i.e,
-	// when the handle column is in a unique index, the real column size may be
-	// smaller than 8 because it is encoded using `EncodeVarint`. Since we don't
-	// know the exact value size now, use 8 as approximation.
-	if c.IsHandle {
-		return 8
-	}
-	histCount := c.TotalRowCount()
-	notNullRatio := 1.0
-	if histCount > 0 {
-		notNullRatio = 1.0 - float64(c.NullCount)/histCount
-	}
-	switch c.Histogram.Tp.GetType() {
-	case mysql.TypeFloat, mysql.TypeDouble, mysql.TypeDuration, mysql.TypeDate, mysql.TypeDatetime, mysql.TypeTimestamp:
-		return 8 * notNullRatio
-	case mysql.TypeTiny, mysql.TypeShort, mysql.TypeInt24, mysql.TypeLong, mysql.TypeLonglong, mysql.TypeYear, mysql.TypeEnum, mysql.TypeBit, mysql.TypeSet:
-		if isKey {
-			return 8 * notNullRatio
-		}
-	}
-	// Keep two decimal place.
-	return math.Round(float64(c.TotColSize)/float64(count)*100) / 100
-}
-
-// AvgColSizeChunkFormat is the average column size of the histogram. These sizes are derived from function `Encode`
-// and `DecodeToChunk`, so we need to update them if those 2 functions are changed.
-func (c *Column) AvgColSizeChunkFormat(count int64) float64 {
-	if count == 0 {
-		return 0
-	}
-	fixedLen := chunk.GetFixedLen(c.Histogram.Tp)
-	if fixedLen != -1 {
-		return float64(fixedLen)
-	}
-	// Keep two decimal place.
-	// Add 8 bytes for unfixed-len type's offsets.
-	// Minus Log2(avgSize) for unfixed-len type LEN.
-	avgSize := float64(c.TotColSize) / float64(count)
-	if avgSize < 1 {
-		return math.Round(avgSize*100)/100 + 8
-	}
-	return math.Round((avgSize-math.Log2(avgSize))*100)/100 + 8
-}
-
-// AvgColSizeListInDisk is the average column size of the histogram. These sizes are derived
-// from `chunk.ListInDisk` so we need to update them if those 2 functions are changed.
-func (c *Column) AvgColSizeListInDisk(count int64) float64 {
-	if count == 0 {
-		return 0
-	}
-	histCount := c.TotalRowCount()
-	notNullRatio := 1.0
-	if histCount > 0 {
-		notNullRatio = 1.0 - float64(c.NullCount)/histCount
-	}
-	size := chunk.GetFixedLen(c.Histogram.Tp)
-	if size != -1 {
-		return float64(size) * notNullRatio
-	}
-	// Keep two decimal place.
-	// Minus Log2(avgSize) for unfixed-len type LEN.
-	avgSize := float64(c.TotColSize) / float64(count)
-	if avgSize < 1 {
-		return math.Round((avgSize)*100) / 100
-	}
-	return math.Round((avgSize-math.Log2(avgSize))*100) / 100
 }
 
 // AppendBucket appends a bucket into `hg`.
@@ -543,15 +465,6 @@ func (hg *Histogram) BetweenRowCount(a, b types.Datum) float64 {
 	return lessCountB - lessCountA
 }
 
-// BetweenRowCount estimates the row count for interval [l, r).
-func (c *Column) BetweenRowCount(sctx sessionctx.Context, l, r types.Datum, lowEncoded, highEncoded []byte) float64 {
-	histBetweenCnt := c.Histogram.BetweenRowCount(l, r)
-	if c.StatsVer <= Version1 {
-		return histBetweenCnt
-	}
-	return float64(c.TopN.BetweenCount(lowEncoded, highEncoded)) + histBetweenCnt
-}
-
 // TotalRowCount returns the total count of this histogram.
 func (hg *Histogram) TotalRowCount() float64 {
 	return hg.notNullCount() + float64(hg.NullCount)
@@ -587,24 +500,6 @@ func (hg *Histogram) mergeBuckets(bucketIdx int) {
 	}
 	hg.Bounds = c
 	hg.Buckets = hg.Buckets[:curBuck]
-}
-
-// GetIncreaseFactor get the increase factor to adjust the final estimated count when the table is modified.
-func (idx *Index) GetIncreaseFactor(realtimeRowCount int64) float64 {
-	columnCount := idx.TotalRowCount()
-	if columnCount == 0 {
-		return 1.0
-	}
-	return float64(realtimeRowCount) / columnCount
-}
-
-// BetweenRowCount estimates the row count for interval [l, r).
-func (idx *Index) BetweenRowCount(l, r types.Datum) float64 {
-	histBetweenCnt := idx.Histogram.BetweenRowCount(l, r)
-	if idx.StatsVer == Version1 {
-		return histBetweenCnt
-	}
-	return float64(idx.TopN.BetweenCount(l.GetBytes(), r.GetBytes())) + histBetweenCnt
 }
 
 // GetIncreaseFactor will return a factor of data increasing after the last analysis.
@@ -1043,686 +938,6 @@ func (e *ErrorRate) Merge(rate *ErrorRate) {
 	e.ErrorTotal += rate.ErrorTotal
 }
 
-// Column represents a column histogram.
-type Column struct {
-	Histogram
-	*CMSketch
-	*TopN
-	*FMSketch
-	PhysicalID int64
-	Count      int64
-	Info       *model.ColumnInfo
-	IsHandle   bool
-	ErrorRate
-	Flag           int64
-	LastAnalyzePos types.Datum
-	StatsVer       int64 // StatsVer is the version of the current stats, used to maintain compatibility
-
-	// StatsLoadedStatus indicates the status of column statistics
-	StatsLoadedStatus
-}
-
-func (c *Column) String() string {
-	return c.Histogram.ToString(0)
-}
-
-// TotalRowCount returns the total count of this column.
-func (c *Column) TotalRowCount() float64 {
-	if c.StatsVer >= Version2 {
-		return c.Histogram.TotalRowCount() + float64(c.TopN.TotalCount())
-	}
-	return c.Histogram.TotalRowCount()
-}
-
-func (c *Column) notNullCount() float64 {
-	if c.StatsVer >= Version2 {
-		return c.Histogram.notNullCount() + float64(c.TopN.TotalCount())
-	}
-	return c.Histogram.notNullCount()
-}
-
-// GetIncreaseFactor get the increase factor to adjust the final estimated count when the table is modified.
-func (c *Column) GetIncreaseFactor(realtimeRowCount int64) float64 {
-	columnCount := c.TotalRowCount()
-	if columnCount == 0 {
-		// avoid dividing by 0
-		return 1.0
-	}
-	return float64(realtimeRowCount) / columnCount
-}
-
-// MemoryUsage returns the total memory usage of Histogram, CMSketch, FMSketch in Column.
-// We ignore the size of other metadata in Column
-func (c *Column) MemoryUsage() CacheItemMemoryUsage {
-	var sum int64
-	columnMemUsage := &ColumnMemUsage{
-		ColumnID: c.Info.ID,
-	}
-	histogramMemUsage := c.Histogram.MemoryUsage()
-	columnMemUsage.HistogramMemUsage = histogramMemUsage
-	sum = histogramMemUsage
-	if c.CMSketch != nil {
-		cmSketchMemUsage := c.CMSketch.MemoryUsage()
-		columnMemUsage.CMSketchMemUsage = cmSketchMemUsage
-		sum += cmSketchMemUsage
-	}
-	if c.TopN != nil {
-		topnMemUsage := c.TopN.MemoryUsage()
-		columnMemUsage.TopNMemUsage = topnMemUsage
-		sum += topnMemUsage
-	}
-	if c.FMSketch != nil {
-		fmSketchMemUsage := c.FMSketch.MemoryUsage()
-		columnMemUsage.FMSketchMemUsage = fmSketchMemUsage
-		sum += fmSketchMemUsage
-	}
-	columnMemUsage.TotalMemUsage = sum
-	return columnMemUsage
-}
-
-// HistogramNeededItems stores the columns/indices whose Histograms need to be loaded from physical kv layer.
-// Currently, we only load index/pk's Histogram from kv automatically. Columns' are loaded by needs.
-var HistogramNeededItems = neededStatsMap{items: map[model.TableItemID]struct{}{}}
-
-// IsInvalid checks if this column is invalid. If this column has histogram but not loaded yet, then we mark it
-// as need histogram.
-func (c *Column) IsInvalid(sctx sessionctx.Context, collPseudo bool) bool {
-	if collPseudo && c.NotAccurate() {
-		return true
-	}
-	if sctx != nil {
-		stmtctx := sctx.GetSessionVars().StmtCtx
-		if stmtctx != nil && stmtctx.StatsLoad.Fallback {
-			return true
-		}
-		if c.IsLoadNeeded() && stmtctx != nil {
-			if stmtctx.StatsLoad.Timeout > 0 {
-				logutil.BgLogger().Warn("Hist for column should already be loaded as sync but not found.",
-					zap.String(strconv.FormatInt(c.Info.ID, 10), c.Info.Name.O))
-			}
-			// In some tests, the c.Info is not set, so we add this check here.
-			if c.Info != nil {
-				HistogramNeededItems.insert(model.TableItemID{TableID: c.PhysicalID, ID: c.Info.ID, IsIndex: false})
-			}
-		}
-	}
-	// In some cases, some statistics in column would be evicted
-	// For example: the cmsketch of the column might be evicted while the histogram and the topn are still exists
-	// In this case, we will think this column as valid due to we can still use the rest of the statistics to do optimize.
-	return c.TotalRowCount() == 0 || (!c.IsEssentialStatsLoaded() && c.Histogram.NDV > 0)
-}
-
-// IsHistNeeded checks if this column needs histogram to be loaded
-func (c *Column) IsHistNeeded(collPseudo bool) bool {
-	return (!collPseudo || !c.NotAccurate()) && c.IsLoadNeeded()
-}
-
-func (c *Column) equalRowCount(sctx sessionctx.Context, val types.Datum, encodedVal []byte, realtimeRowCount int64) (float64, error) {
-	if val.IsNull() {
-		return float64(c.NullCount), nil
-	}
-	if c.StatsVer < Version2 {
-		// All the values are null.
-		if c.Histogram.Bounds.NumRows() == 0 {
-			return 0.0, nil
-		}
-		if c.Histogram.NDV > 0 && c.outOfRange(val) {
-			return outOfRangeEQSelectivity(c.Histogram.NDV, realtimeRowCount, int64(c.TotalRowCount())) * c.TotalRowCount(), nil
-		}
-		if c.CMSketch != nil {
-			count, err := queryValue(sctx.GetSessionVars().StmtCtx, c.CMSketch, c.TopN, val)
-			return float64(count), errors.Trace(err)
-		}
-		histRowCount, _ := c.Histogram.equalRowCount(val, false)
-		return histRowCount, nil
-	}
-
-	// Stats version == 2
-	// All the values are null.
-	if c.Histogram.Bounds.NumRows() == 0 && c.TopN.Num() == 0 {
-		return 0, nil
-	}
-	// 1. try to find this value in TopN
-	if c.TopN != nil {
-		rowcount, ok := c.QueryTopN(encodedVal)
-		if ok {
-			return float64(rowcount), nil
-		}
-	}
-	// 2. try to find this value in bucket.Repeat(the last value in every bucket)
-	histCnt, matched := c.Histogram.equalRowCount(val, true)
-	if matched {
-		return histCnt, nil
-	}
-	// 3. use uniform distribution assumption for the rest (even when this value is not covered by the range of stats)
-	histNDV := float64(c.Histogram.NDV - int64(c.TopN.Num()))
-	if histNDV <= 0 {
-		return 0, nil
-	}
-	return c.Histogram.notNullCount() / histNDV, nil
-}
-
-// GetColumnRowCount estimates the row count by a slice of Range.
-func (c *Column) GetColumnRowCount(sctx sessionctx.Context, ranges []*ranger.Range, realtimeRowCount int64, pkIsHandle bool) (float64, error) {
-	sc := sctx.GetSessionVars().StmtCtx
-	var rowCount float64
-	for _, rg := range ranges {
-		highVal := *rg.HighVal[0].Clone()
-		lowVal := *rg.LowVal[0].Clone()
-		if highVal.Kind() == types.KindString {
-			highVal.SetBytes(collate.GetCollator(highVal.Collation()).Key(highVal.GetString()))
-		}
-		if lowVal.Kind() == types.KindString {
-			lowVal.SetBytes(collate.GetCollator(lowVal.Collation()).Key(lowVal.GetString()))
-		}
-		cmp, err := lowVal.Compare(sc, &highVal, collate.GetBinaryCollator())
-		if err != nil {
-			return 0, errors.Trace(err)
-		}
-		lowEncoded, err := codec.EncodeKey(sc, nil, lowVal)
-		if err != nil {
-			return 0, err
-		}
-		highEncoded, err := codec.EncodeKey(sc, nil, highVal)
-		if err != nil {
-			return 0, err
-		}
-		if cmp == 0 {
-			// case 1: it's a point
-			if !rg.LowExclude && !rg.HighExclude {
-				// In this case, the row count is at most 1.
-				if pkIsHandle {
-					rowCount += 1
-					continue
-				}
-				var cnt float64
-				cnt, err = c.equalRowCount(sctx, lowVal, lowEncoded, realtimeRowCount)
-				if err != nil {
-					return 0, errors.Trace(err)
-				}
-				// If the current table row count has changed, we should scale the row count accordingly.
-				cnt *= c.GetIncreaseFactor(realtimeRowCount)
-				rowCount += cnt
-			}
-			continue
-		}
-		// In stats ver 1, we use CM Sketch to estimate row count for point condition, which is more accurate.
-		// So for the small range, we convert it to points.
-		if c.StatsVer < 2 {
-			rangeVals := enumRangeValues(lowVal, highVal, rg.LowExclude, rg.HighExclude)
-
-			// case 2: it's a small range && using ver1 stats
-			if rangeVals != nil {
-				for _, val := range rangeVals {
-					cnt, err := c.equalRowCount(sctx, val, lowEncoded, realtimeRowCount)
-					if err != nil {
-						return 0, err
-					}
-					// If the current table row count has changed, we should scale the row count accordingly.
-					cnt *= c.GetIncreaseFactor(realtimeRowCount)
-					rowCount += cnt
-				}
-
-				continue
-			}
-		}
-
-		// case 3: it's an interval
-		cnt := c.BetweenRowCount(sctx, lowVal, highVal, lowEncoded, highEncoded)
-		// `betweenRowCount` returns count for [l, h) range, we adjust cnt for boundaries here.
-		// Note that, `cnt` does not include null values, we need specially handle cases
-		//   where null is the lower bound.
-		// And because we use (2, MaxValue] to represent expressions like a > 2 and use [MinNotNull, 3) to represent
-		//   expressions like b < 3, we need to exclude the special values.
-		if rg.LowExclude && !lowVal.IsNull() && lowVal.Kind() != types.KindMaxValue && lowVal.Kind() != types.KindMinNotNull {
-			lowCnt, err := c.equalRowCount(sctx, lowVal, lowEncoded, realtimeRowCount)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			cnt -= lowCnt
-			cnt = mathutil.Clamp(cnt, 0, c.notNullCount())
-		}
-		if !rg.LowExclude && lowVal.IsNull() {
-			cnt += float64(c.NullCount)
-		}
-		if !rg.HighExclude && highVal.Kind() != types.KindMaxValue && highVal.Kind() != types.KindMinNotNull {
-			highCnt, err := c.equalRowCount(sctx, highVal, highEncoded, realtimeRowCount)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			cnt += highCnt
-		}
-
-		cnt = mathutil.Clamp(cnt, 0, c.TotalRowCount())
-
-		// If the current table row count has changed, we should scale the row count accordingly.
-		cnt *= c.GetIncreaseFactor(realtimeRowCount)
-
-		// handling the out-of-range part
-		if (c.outOfRange(lowVal) && !lowVal.IsNull()) || c.outOfRange(highVal) {
-			increaseCount := realtimeRowCount - int64(c.TotalRowCount())
-			if increaseCount < 0 {
-				increaseCount = 0
-			}
-			cnt += c.Histogram.outOfRangeRowCount(&lowVal, &highVal, increaseCount)
-		}
-
-		rowCount += cnt
-	}
-	rowCount = mathutil.Clamp(rowCount, 0, float64(realtimeRowCount))
-	return rowCount, nil
-}
-
-// ItemID implements TableCacheItem
-func (c *Column) ItemID() int64 {
-	return c.Info.ID
-}
-
-// DropEvicted implements TableCacheItem
-// DropEvicted drops evicted structures
-func (c *Column) DropEvicted() {
-	if !c.statsInitialized {
-		return
-	}
-	switch c.evictedStatus {
-	case allLoaded:
-		if c.CMSketch != nil && c.StatsVer < Version2 {
-			c.dropCMS()
-			return
-		}
-		// For stats version2, there is no cms thus we directly drop topn
-		c.dropTopN()
-		return
-	case onlyCmsEvicted:
-		c.dropTopN()
-		return
-	default:
-		return
-	}
-}
-
-func (c *Column) dropCMS() {
-	c.CMSketch = nil
-	c.evictedStatus = onlyCmsEvicted
-}
-
-func (c *Column) dropTopN() {
-	originTopNNum := int64(c.TopN.Num())
-	c.TopN = nil
-	if len(c.Histogram.Buckets) == 0 && originTopNNum >= c.Histogram.NDV {
-		// This indicates column has topn instead of histogram
-		c.evictedStatus = allEvicted
-	} else {
-		c.evictedStatus = onlyHistRemained
-	}
-}
-
-// IsAllEvicted indicates whether all stats evicted
-func (c *Column) IsAllEvicted() bool {
-	return c.statsInitialized && c.evictedStatus >= allEvicted
-}
-
-func (c *Column) getEvictedStatus() int {
-	return c.evictedStatus
-}
-
-func (c *Column) isStatsInitialized() bool {
-	return c.statsInitialized
-}
-
-func (c *Column) statsVer() int64 {
-	return c.StatsVer
-}
-
-func (c *Column) isCMSExist() bool {
-	return c.CMSketch != nil
-}
-
-// Index represents an index histogram.
-type Index struct {
-	Histogram
-	*CMSketch
-	*TopN
-	FMSketch *FMSketch
-	ErrorRate
-	StatsVer       int64 // StatsVer is the version of the current stats, used to maintain compatibility
-	Info           *model.IndexInfo
-	Flag           int64
-	LastAnalyzePos types.Datum
-	PhysicalID     int64
-	StatsLoadedStatus
-}
-
-// ItemID implements TableCacheItem
-func (idx *Index) ItemID() int64 {
-	return idx.Info.ID
-}
-
-// IsAllEvicted indicates whether all stats evicted
-func (idx *Index) IsAllEvicted() bool {
-	return idx.statsInitialized && idx.evictedStatus >= allEvicted
-}
-
-func (idx *Index) dropCMS() {
-	idx.CMSketch = nil
-	idx.evictedStatus = onlyCmsEvicted
-}
-
-func (idx *Index) dropTopN() {
-	originTopNNum := int64(idx.TopN.Num())
-	idx.TopN = nil
-	if len(idx.Histogram.Buckets) == 0 && originTopNNum >= idx.Histogram.NDV {
-		// This indicates index has topn instead of histogram
-		idx.evictedStatus = allEvicted
-	} else {
-		idx.evictedStatus = onlyHistRemained
-	}
-}
-
-func (idx *Index) getEvictedStatus() int {
-	return idx.evictedStatus
-}
-
-func (idx *Index) isStatsInitialized() bool {
-	return idx.statsInitialized
-}
-
-func (idx *Index) statsVer() int64 {
-	return idx.StatsVer
-}
-
-func (idx *Index) isCMSExist() bool {
-	return idx.CMSketch != nil
-}
-
-// IsEvicted returns whether index statistics got evicted
-func (idx *Index) IsEvicted() bool {
-	return idx.evictedStatus != allLoaded
-}
-
-func (idx *Index) String() string {
-	return idx.Histogram.ToString(len(idx.Info.Columns))
-}
-
-// TotalRowCount returns the total count of this index.
-func (idx *Index) TotalRowCount() float64 {
-	idx.checkStats()
-	if idx.StatsVer >= Version2 {
-		return idx.Histogram.TotalRowCount() + float64(idx.TopN.TotalCount())
-	}
-	return idx.Histogram.TotalRowCount()
-}
-
-// IsInvalid checks if this index is invalid.
-func (idx *Index) IsInvalid(collPseudo bool) bool {
-	if !collPseudo {
-		idx.checkStats()
-	}
-	return (collPseudo && idx.NotAccurate()) || idx.TotalRowCount() == 0
-}
-
-// EvictAllStats evicts all stats
-// Note that this function is only used for test
-func (idx *Index) EvictAllStats() {
-	idx.Buckets = nil
-	idx.CMSketch = nil
-	idx.TopN = nil
-	idx.StatsLoadedStatus.evictedStatus = allEvicted
-}
-
-// MemoryUsage returns the total memory usage of a Histogram and CMSketch in Index.
-// We ignore the size of other metadata in Index.
-func (idx *Index) MemoryUsage() CacheItemMemoryUsage {
-	var sum int64
-	indexMemUsage := &IndexMemUsage{
-		IndexID: idx.Info.ID,
-	}
-	histMemUsage := idx.Histogram.MemoryUsage()
-	indexMemUsage.HistogramMemUsage = histMemUsage
-	sum = histMemUsage
-	if idx.CMSketch != nil {
-		cmSketchMemUsage := idx.CMSketch.MemoryUsage()
-		indexMemUsage.CMSketchMemUsage = cmSketchMemUsage
-		sum += cmSketchMemUsage
-	}
-	if idx.TopN != nil {
-		topnMemUsage := idx.TopN.MemoryUsage()
-		indexMemUsage.TopNMemUsage = topnMemUsage
-		sum += topnMemUsage
-	}
-	indexMemUsage.TotalMemUsage = sum
-	return indexMemUsage
-}
-
-var nullKeyBytes, _ = codec.EncodeKey(nil, nil, types.NewDatum(nil))
-
-func (idx *Index) equalRowCount(b []byte, realtimeRowCount int64) float64 {
-	if len(idx.Info.Columns) == 1 {
-		if bytes.Equal(b, nullKeyBytes) {
-			return float64(idx.NullCount)
-		}
-	}
-	val := types.NewBytesDatum(b)
-	if idx.StatsVer < Version2 {
-		if idx.NDV > 0 && idx.outOfRange(val) {
-			return outOfRangeEQSelectivity(idx.NDV, realtimeRowCount, int64(idx.TotalRowCount())) * idx.TotalRowCount()
-		}
-		if idx.CMSketch != nil {
-			return float64(idx.QueryBytes(b))
-		}
-		histRowCount, _ := idx.Histogram.equalRowCount(val, false)
-		return histRowCount
-	}
-	// stats version == 2
-	// 1. try to find this value in TopN
-	count, found := idx.TopN.QueryTopN(b)
-	if found {
-		return float64(count)
-	}
-	// 2. try to find this value in bucket.Repeat(the last value in every bucket)
-	histCnt, matched := idx.Histogram.equalRowCount(val, true)
-	if matched {
-		return histCnt
-	}
-	// 3. use uniform distribution assumption for the rest (even when this value is not covered by the range of stats)
-	histNDV := float64(idx.Histogram.NDV - int64(idx.TopN.Num()))
-	if histNDV <= 0 {
-		return 0
-	}
-	return idx.Histogram.notNullCount() / histNDV
-}
-
-// QueryBytes is used to query the count of specified bytes.
-func (idx *Index) QueryBytes(d []byte) uint64 {
-	idx.checkStats()
-	h1, h2 := murmur3.Sum128(d)
-	if count, ok := idx.TopN.QueryTopN(d); ok {
-		return count
-	}
-	if idx.CMSketch != nil {
-		return idx.queryHashValue(h1, h2)
-	}
-	v, _ := idx.Histogram.equalRowCount(types.NewBytesDatum(d), idx.StatsVer >= Version2)
-	return uint64(v)
-}
-
-// GetRowCount returns the row count of the given ranges.
-// It uses the modifyCount to adjust the influence of modifications on the table.
-func (idx *Index) GetRowCount(sctx sessionctx.Context, coll *HistColl, indexRanges []*ranger.Range, realtimeRowCount int64) (float64, error) {
-	idx.checkStats()
-	sc := sctx.GetSessionVars().StmtCtx
-	totalCount := float64(0)
-	isSingleCol := len(idx.Info.Columns) == 1
-	for _, indexRange := range indexRanges {
-		lb, err := codec.EncodeKey(sc, nil, indexRange.LowVal...)
-		if err != nil {
-			return 0, err
-		}
-		rb, err := codec.EncodeKey(sc, nil, indexRange.HighVal...)
-		if err != nil {
-			return 0, err
-		}
-		fullLen := len(indexRange.LowVal) == len(indexRange.HighVal) && len(indexRange.LowVal) == len(idx.Info.Columns)
-		if bytes.Equal(lb, rb) {
-			// case 1: it's a point
-			if indexRange.LowExclude || indexRange.HighExclude {
-				continue
-			}
-			if fullLen {
-				// At most 1 in this case.
-				if idx.Info.Unique {
-					totalCount += 1
-					continue
-				}
-				count := idx.equalRowCount(lb, realtimeRowCount)
-				// If the current table row count has changed, we should scale the row count accordingly.
-				count *= idx.GetIncreaseFactor(realtimeRowCount)
-				totalCount += count
-				continue
-			}
-		}
-
-		// case 2: it's an interval
-		// The final interval is [low, high)
-		if indexRange.LowExclude {
-			lb = kv.Key(lb).PrefixNext()
-		}
-		if !indexRange.HighExclude {
-			rb = kv.Key(rb).PrefixNext()
-		}
-		l := types.NewBytesDatum(lb)
-		r := types.NewBytesDatum(rb)
-		lowIsNull := bytes.Equal(lb, nullKeyBytes)
-		if isSingleCol && lowIsNull {
-			totalCount += float64(idx.NullCount)
-		}
-		expBackoffSuccess := false
-		// Due to the limitation of calcFraction and convertDatumToScalar, the histogram actually won't estimate anything.
-		// If the first column's range is point.
-		if rangePosition := GetOrdinalOfRangeCond(sc, indexRange); rangePosition > 0 && idx.StatsVer >= Version2 && coll != nil {
-			var expBackoffSel float64
-			expBackoffSel, expBackoffSuccess, err = idx.expBackoffEstimation(sctx, coll, indexRange)
-			if err != nil {
-				return 0, err
-			}
-			if expBackoffSuccess {
-				expBackoffCnt := expBackoffSel * idx.TotalRowCount()
-
-				upperLimit := expBackoffCnt
-				// Use the multi-column stats to calculate the max possible row count of [l, r)
-				if idx.Len() > 0 {
-					_, lowerBkt, _, _ := idx.locateBucket(l)
-					_, upperBkt, _, _ := idx.locateBucket(r)
-					// Use Count of the Bucket before l as the lower bound.
-					preCount := float64(0)
-					if lowerBkt > 0 {
-						preCount = float64(idx.Buckets[lowerBkt-1].Count)
-					}
-					// Use Count of the Bucket where r exists as the upper bound.
-					upperCnt := float64(idx.Buckets[upperBkt].Count)
-					upperLimit = upperCnt - preCount
-					upperLimit += float64(idx.TopN.BetweenCount(lb, rb))
-				}
-
-				// If the result of exponential backoff strategy is larger than the result from multi-column stats,
-				// 	use the upper limit from multi-column histogram instead.
-				if expBackoffCnt > upperLimit {
-					expBackoffCnt = upperLimit
-				}
-				totalCount += expBackoffCnt
-			}
-		}
-		if !expBackoffSuccess {
-			totalCount += idx.BetweenRowCount(l, r)
-		}
-
-		// If the current table row count has changed, we should scale the row count accordingly.
-		totalCount *= idx.GetIncreaseFactor(realtimeRowCount)
-
-		// handling the out-of-range part
-		if (idx.outOfRange(l) && !(isSingleCol && lowIsNull)) || idx.outOfRange(r) {
-			increaseCount := realtimeRowCount - int64(idx.TotalRowCount())
-			if increaseCount < 0 {
-				increaseCount = 0
-			}
-			totalCount += idx.Histogram.outOfRangeRowCount(&l, &r, increaseCount)
-		}
-	}
-	totalCount = mathutil.Clamp(totalCount, 0, float64(realtimeRowCount))
-	return totalCount, nil
-}
-
-// expBackoffEstimation estimate the multi-col cases following the Exponential Backoff. See comment below for details.
-func (idx *Index) expBackoffEstimation(sctx sessionctx.Context, coll *HistColl, indexRange *ranger.Range) (float64, bool, error) {
-	tmpRan := []*ranger.Range{
-		{
-			LowVal:    make([]types.Datum, 1),
-			HighVal:   make([]types.Datum, 1),
-			Collators: make([]collate.Collator, 1),
-		},
-	}
-	colsIDs := coll.Idx2ColumnIDs[idx.ID]
-	singleColumnEstResults := make([]float64, 0, len(indexRange.LowVal))
-	// The following codes uses Exponential Backoff to reduce the impact of independent assumption. It works like:
-	//   1. Calc the selectivity of each column.
-	//   2. Sort them and choose the first 4 most selective filter and the corresponding selectivity is sel_1, sel_2, sel_3, sel_4 where i < j => sel_i < sel_j.
-	//   3. The final selectivity would be sel_1 * sel_2^{1/2} * sel_3^{1/4} * sel_4^{1/8}.
-	// This calculation reduced the independence assumption and can work well better than it.
-	for i := 0; i < len(indexRange.LowVal); i++ {
-		tmpRan[0].LowVal[0] = indexRange.LowVal[i]
-		tmpRan[0].HighVal[0] = indexRange.HighVal[i]
-		tmpRan[0].Collators[0] = indexRange.Collators[0]
-		if i == len(indexRange.LowVal)-1 {
-			tmpRan[0].LowExclude = indexRange.LowExclude
-			tmpRan[0].HighExclude = indexRange.HighExclude
-		}
-		colID := colsIDs[i]
-		var (
-			count float64
-			err   error
-		)
-		if anotherIdxID, ok := coll.ColID2IdxID[colID]; ok && anotherIdxID != idx.ID {
-			count, err = coll.GetRowCountByIndexRanges(sctx, anotherIdxID, tmpRan)
-		} else if col, ok := coll.Columns[colID]; ok && !col.IsInvalid(sctx, coll.Pseudo) {
-			count, err = coll.GetRowCountByColumnRanges(sctx, colID, tmpRan)
-		} else {
-			continue
-		}
-		if err != nil {
-			return 0, false, err
-		}
-		singleColumnEstResults = append(singleColumnEstResults, count)
-	}
-	// Sort them.
-	slices.Sort(singleColumnEstResults)
-	l := len(singleColumnEstResults)
-	// Convert the first 4 to selectivity results.
-	for i := 0; i < l && i < 4; i++ {
-		singleColumnEstResults[i] = singleColumnEstResults[i] / float64(coll.Count)
-	}
-	failpoint.Inject("cleanEstResults", func() {
-		singleColumnEstResults = singleColumnEstResults[:0]
-		l = 0
-	})
-	if l == 1 {
-		return singleColumnEstResults[0], true, nil
-	} else if l == 2 {
-		return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]), true, nil
-	} else if l == 3 {
-		return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]) * math.Sqrt(math.Sqrt(singleColumnEstResults[2])), true, nil
-	} else if l == 0 {
-		return 0, false, nil
-	}
-	return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]) * math.Sqrt(math.Sqrt(singleColumnEstResults[2])) * math.Sqrt(math.Sqrt(math.Sqrt(singleColumnEstResults[3]))), true, nil
-}
-
-func (idx *Index) checkStats() {
-	if idx.IsFullLoad() {
-		return
-	}
-	HistogramNeededItems.insert(model.TableItemID{TableID: idx.PhysicalID, ID: idx.Info.ID, IsIndex: true})
-}
-
 type countByRangeFunc = func(sessionctx.Context, int64, []*ranger.Range) (float64, error)
 
 // newHistogramBySelectivity fulfills the content of new histogram by the given selectivity result.
@@ -1766,54 +981,6 @@ func newHistogramBySelectivity(sctx sessionctx.Context, histID int64, oldHist, n
 		}
 	}
 	return nil
-}
-
-func (idx *Index) newIndexBySelectivity(sc *stmtctx.StatementContext, statsNode *StatsNode) (*Index, error) {
-	var (
-		ranLowEncode, ranHighEncode []byte
-		err                         error
-	)
-	newIndexHist := &Index{Info: idx.Info, StatsVer: idx.StatsVer, CMSketch: idx.CMSketch, PhysicalID: idx.PhysicalID}
-	newIndexHist.Histogram = *NewHistogram(idx.ID, int64(float64(idx.NDV)*statsNode.Selectivity), 0, 0, types.NewFieldType(mysql.TypeBlob), chunk.InitialCapacity, 0)
-
-	lowBucketIdx, highBucketIdx := 0, 0
-	var totCnt int64
-
-	// Bucket bound of index is encoded one, so we need to decode it if we want to calculate the fraction accurately.
-	// TODO: enhance its calculation.
-	// Now just remove the bucket that no range fell in.
-	for _, ran := range statsNode.Ranges {
-		lowBucketIdx = highBucketIdx
-		ranLowEncode, ranHighEncode, err = ran.Encode(sc, ranLowEncode, ranHighEncode)
-		if err != nil {
-			return nil, err
-		}
-		for ; highBucketIdx < idx.Len(); highBucketIdx++ {
-			// Encoded value can only go to its next quickly. So ranHighEncode is actually range.HighVal's PrefixNext value.
-			// So the Bound should also go to its PrefixNext.
-			bucketLowerEncoded := idx.Bounds.GetRow(highBucketIdx * 2).GetBytes(0)
-			if bytes.Compare(ranHighEncode, kv.Key(bucketLowerEncoded).PrefixNext()) < 0 {
-				break
-			}
-		}
-		for ; lowBucketIdx < highBucketIdx; lowBucketIdx++ {
-			bucketUpperEncoded := idx.Bounds.GetRow(lowBucketIdx*2 + 1).GetBytes(0)
-			if bytes.Compare(ranLowEncode, bucketUpperEncoded) <= 0 {
-				break
-			}
-		}
-		if lowBucketIdx >= idx.Len() {
-			break
-		}
-		for i := lowBucketIdx; i < highBucketIdx; i++ {
-			newIndexHist.Bounds.AppendRow(idx.Bounds.GetRow(i * 2))
-			newIndexHist.Bounds.AppendRow(idx.Bounds.GetRow(i*2 + 1))
-			totCnt += idx.bucketCount(i)
-			newIndexHist.Buckets = append(newIndexHist.Buckets, Bucket{Repeat: idx.Buckets[i].Repeat, Count: totCnt})
-			newIndexHist.scalars = append(newIndexHist.scalars, idx.scalars[i])
-		}
-	}
-	return newIndexHist, nil
 }
 
 // NewHistCollBySelectivity creates new HistColl by the given statsNodes.
@@ -1896,25 +1063,6 @@ func (coll *HistColl) NewHistCollBySelectivity(sctx sessionctx.Context, statsNod
 		}
 	}
 	return newColl
-}
-
-func (idx *Index) outOfRange(val types.Datum) bool {
-	if !idx.Histogram.outOfRange(val) {
-		return false
-	}
-	if idx.Histogram.Len() > 0 && matchPrefix(idx.Bounds.GetRow(0), 0, &val) {
-		return false
-	}
-	return true
-}
-
-// matchPrefix checks whether ad is the prefix of value
-func matchPrefix(row chunk.Row, colIdx int, ad *types.Datum) bool {
-	switch ad.Kind() {
-	case types.KindString, types.KindBytes, types.KindBinaryLiteral, types.KindMysqlBit:
-		return strings.HasPrefix(row.GetString(colIdx), ad.GetString())
-	}
-	return false
 }
 
 type dataCnt struct {
